lee 修改购物车居中

+++ MySuperApp/Class/UI/mainpage/branddetail/ProductContentViewController.m
@@ -81,7 +81,7 @@
     newfvc.params = self.params;
     newfvc.orderStr = @"";
     newfvc.key = @"";
-    newfvc.strcurrentpage = [NSString stringWithFormat:@"%d",current];
+    newfvc.strcurrentpage = [NSString stringWithFormat:@"%ld",(long)current];
     newfvc.strperpage = @"10";
     [self.navigationController pushViewController:newfvc animated:YES];
 }
